api/pyproject.toml
[project]
name = "list-ingestor-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai==1.75.0",
    "psycopg[binary]==3.2.4",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "flask==3.1.0",
    "flask-sqlalchemy==3.1.1",
    "werkzeug==3.1.3",
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "pillow==11.1.0",
    "rapidfuzz==3.12.1",
    "numpy==2.2.3",
    "pydantic==2.11.1",
    "anthropic==0.49.0",
]
